feat: better user authentication
Rather than checking if a 'authenticatedUserID' value exists in the
session, we create a middleware that checks if the 'authenticatedUserID'
exists in the users database also.

This removes the potential problem of a user still being logged in even
after their account has been deleted as the session only updates in like
12 hours, but now due to the authentication being mostly done by
checking in database through a middleware, it will be updated on every
request.

The middleware checks in the database if the 'authenticatedUserID' value
exists in the database and updates the request context with that value.
Now, the request context is used to check if a user is authenticated or
not in the 'isAuthenticated()' function.